+++ library/src/nvidia_detail/hipsolver.cpp
@@ -89,6 +89,49 @@ catch(...)
     return hipsolver::exception2hip_status();
 }
 
+hipsolverStatus_t hipsolverSetDeterministicMode(hipsolverHandle_t            handle,
+                                                hipsolverDeterministicMode_t mode)
+try
+{
+#if(CUDART_VERSION >= 12020)
+    if(!handle)
+        return HIPSOLVER_STATUS_NOT_INITIALIZED;
+
+    return hipsolver::cuda2hip_status(cusolverSetDeterministicMode(
+        (cusolverDnHandle_t)handle, hipsolver::hip2cuda_deterministic(mode)));
+#else
+    return HIPSOLVER_STATUS_NOT_SUPPORTED;
+#endif
+}
+catch(...)
+{
+    return hipsolver::exception2hip_status();
+}
+
+hipsolverStatus_t hipsolverGetDeterministicMode(hipsolverHandle_t             handle,
+                                                hipsolverDeterministicMode_t* mode)
+try
+{
+#if(CUDART_VERSION >= 12020)
+    if(!handle)
+        return HIPSOLVER_STATUS_NOT_INITIALIZED;
+    if(!mode)
+        return HIPSOLVER_STATUS_INVALID_VALUE;
+
+    cusolverDeterministicMode_t dmode;
+    CHECK_CUSOLVER_ERROR(cusolverGetDeterministicMode((cusolverDnHandle_t)handle, &dmode));
+    *mode = hipsolver::cuda2hip_deterministic(dmode);
+
+    return HIPSOLVER_STATUS_SUCCESS;
+#else
+    return HIPSOLVER_STATUS_NOT_SUPPORTED;
+#endif
+}
+catch(...)
+{
+    return hipsolver::exception2hip_status();
+}
+
 /******************** GESVDJ PARAMS ********************/
 hipsolverStatus_t hipsolverCreateGesvdjInfo(hipsolverGesvdjInfo_t* info)
 try